backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "AI Desktop Client Backend"
authors = ["MassMash"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.133.1"}
uvicorn = "^0.41.0"
python-multipart = "^0.0.22"
python-dotenv = "^1.2.1"
pydantic = "^2.12.5"
httpx = "^0.28.1"
pdfplumber = "^0.11.9"
python-docx = "^1.2.0"
openai = "^2.24.0"
google-generativeai = "^0.8.6"
anthropic = "^0.84.0"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
